fix level1 messing up initial layout, and rework tests
