+++ test/python/compiler/test_compiler.py
@@ -210,26 +210,6 @@ class TestCompiler(QiskitTestCase):
         rfalse = backend.run(qrfalse).result()
         self.assertEqual(rtrue.get_counts(), rfalse.get_counts())
 
-    def test_compile_with_initial_layout(self):
-        """Test compile with an initial layout.
-        Regression test for #1711
-        """
-        qr = QuantumRegister(3)
-        cr = ClassicalRegister(3)
-        qc = QuantumCircuit(qr, cr)
-        qc.cx(qr[2], qr[1])
-        qc.cx(qr[2], qr[0])
-        initial_layout = {0: qr[1], 2: qr[0], 15: qr[2]}
-        backend = FakeRueschlikon()
-        qc_b = transpile(qc, backend, seed_transpiler=42, initial_layout=initial_layout)
-        qobj = assemble(qc_b)
-
-        compiled_ops = qobj.experiments[0].instructions
-        for operation in compiled_ops:
-            if operation.name == 'cx':
-                self.assertIn(operation.qubits, backend.configuration().coupling_map)
-                self.assertIn(operation.qubits, [[15, 0], [15, 2]])
-
     def test_mapper_overoptimization(self):
         """Check mapper overoptimization.
